Moved the router element to utils / Router.
