+++ src/App.js
@@ -3,19 +3,21 @@ import styled from 'styled-components'
 
 // import { RouterProvider } from 'react-router-dom'
 // import { router } from './utils/Router'
-import {
-  RouterProvider,
-  createBrowserRouter,
-  createRoutesFromElements,
-  Route,
-  Outlet,
-  useParams,
-} from 'react-router-dom'
+// import {
+//   RouterProvider,
+//   createBrowserRouter,
+//   createRoutesFromElements,
+//   Route,
+//   Outlet,
+//   useParams,
+// } from 'react-router-dom'
 
 import { Grid, SidePanel } from '@digicatapult/ui-component-library'
 import { Context } from './utils/Context'
 
-import Certificates from './pages/Certificates'
+// import Certificates from './pages/Certificates'
+
+import Routes from './utils/Router'
 
 export const personas = [
   {
@@ -54,48 +56,48 @@ export const personas = [
   },
 ]
 
-function Certificate() {
-  let { id } = useParams()
-  return (
-    <>
-      <h3>CertIndex: {id}</h3>
-    </>
-  )
-}
+// function Certificate() {
+//   let { id } = useParams()
+//   return (
+//     <>
+//       <h3>CertIndex: {id}</h3>
+//     </>
+//   )
+// }
 
-function CertificateCo2Embedder() {
-  let { id } = useParams()
-
-  const {
-    current,
-    currentId,
-    currentCommitmentSalt,
-    currentEnergyConsumedWh,
-    currentProductionStartTime,
-    currentProductionEndTime,
-  } = React.useContext(Context)
-
-  return (
-    <>
-      <h3>Cert Embed Id: {id}</h3>
-      <hr />
-      {current == 'heidi' && <>Switch2Emma</>}
-      {current == 'emma' && (
-        <code>
-          cur id: {currentId} <br />
-          currentCommitmentSalt: {currentCommitmentSalt} <br />
-          currentEnergyConsumedWh: {currentEnergyConsumedWh} <br />
-          currentProductionStartTime: {currentProductionStartTime} <br />
-          currentProductionEndTime: {currentProductionEndTime} <br />
-          TODO: embed the co2 data w/ <br />
-          GET /v1/certificate ( get the latest that matches the above ) <br />
-          POST v1/certificate/$emma_local_id <br />
-          POST v1/certificate/$emma_local_id/issuance
-        </code>
-      )}
-    </>
-  )
-}
+// function CertificateCo2Embedder() {
+//   let { id } = useParams()
+//
+//   const {
+//     current,
+//     currentId,
+//     currentCommitmentSalt,
+//     currentEnergyConsumedWh,
+//     currentProductionStartTime,
+//     currentProductionEndTime,
+//   } = React.useContext(Context)
+//
+//   return (
+//     <>
+//       <h3>Cert Embed Id: {id}</h3>
+//       <hr />
+//       {current == 'heidi' && <>Switch2Emma</>}
+//       {current == 'emma' && (
+//         <code>
+//           cur id: {currentId} <br />
+//           currentCommitmentSalt: {currentCommitmentSalt} <br />
+//           currentEnergyConsumedWh: {currentEnergyConsumedWh} <br />
+//           currentProductionStartTime: {currentProductionStartTime} <br />
+//           currentProductionEndTime: {currentProductionEndTime} <br />
+//           TODO: embed the co2 data w/ <br />
+//           GET /v1/certificate ( get the latest that matches the above ) <br />
+//           POST v1/certificate/$emma_local_id <br />
+//           POST v1/certificate/$emma_local_id/issuance
+//         </code>
+//       )}
+//     </>
+//   )
+// }
 
 export default function App() {
   const {
@@ -157,9 +159,8 @@ export default function App() {
       {/* {(currentId != '' || currentCommitmentSalt != '') && ( */}
       {/* <>{currentCommitmentSalt}</> */}
       {/* )} */}
-
       {/* <RouterProvider router={router} /> */}
-      <RouterProvider
+      {/* <RouterProvider
         router={createBrowserRouter(
           createRoutesFromElements(
             <>
@@ -177,8 +178,9 @@ export default function App() {
             </>
           )
         )}
-      />
+      /> */}
       {/* <BrowserRouter /> */}
+      <Routes />
     </FullScreenGrid>
   )
 }
